Baseline 6.0 development squashed from current wip/6.0 as of 2018-11-27

+++ hibernate-core/src/main/java/org/hibernate/boot/model/source/spi/SizeSource.java
@@ -15,7 +15,7 @@ public interface SizeSource {
 	 *
 	 * @return The length, or {@code null} if not defined.
 	 */
-	Integer getLength();
+	Long getLength();
 
 	/**
 	 * The specified precision.  Will return {@code null} if none was specified.
